added example data and significant test coverage
Wow, this unittest coverage game is kind of addicting...

Anyways, added an example input CSV file and output txt file for examples of
what this tool does. I played around with looking at code coverage, then
decided to go overboard in unit test coverage. Maybe it's excessive, maybe
not, it's my first time doing this. Got the coverage number up to like 95 so
I think it's good to call it for now